Create UpdatedNewTeamsBackEnd.py
Needed for this update it will be a another csv file called 'schedule.csv', also it prints out to the csv file but does not sort them in order. Im work on a new methond to def sort_schedule() to sort them in order. Other than that it prints to the csv file. Also the display_schedule isn't working, im still figuring it out. Also its not really needed but its more for just testing in the back end so i might end up removing it.